+++ C+lan/C+lan/ScreenInput.cpp
@@ -11,7 +11,7 @@
 Ctlan::PublicSystem::Vector2 Ctlan::PublicSystem::ScreenInput::GetScreenPosition(Vector2 _paramPosition)
 {
     // スクリーン座標を代入する変数（引数の値を初期値として代入）
-    POINT ScreenPosition = {_paramPosition.x, _paramPosition.y};
+    POINT ScreenPosition = {(LONG)_paramPosition.x, (LONG)_paramPosition.y};
 
     // スクリーン座標を取得
     ScreenToClient(PrivateSystem::Application::GetWindow(), &ScreenPosition);
@@ -20,5 +20,5 @@ Ctlan::PublicSystem::Vector2 Ctlan::PublicSystem::ScreenInput::GetScreenPosition
     // Clientはアプリの画面（ScreenInputのScreenに値するもの）を指している。
 
     // 値をVector2に変換して返す
-    return Vector2(ScreenPosition.x, ScreenPosition.y);
+    return Vector2((float)ScreenPosition.x, (float)ScreenPosition.y);
 }
